requirements.txt
discord.py
open-interpreter
psutil
cryptography
gTTS
requests
flask